tests/requirements.txt
# Python dependencies for Ghostty Android tests

# Image processing (optional, fallback if ImageMagick not available)
Pillow>=10.0.0
numpy>=1.24.0

# Test framework utilities
pytest>=7.4.0
pytest-xdist>=3.3.0  # For parallel test execution

# Code quality
black>=23.0.0
mypy>=1.5.0
